refactor: 테스트 오타 수정
